requirements.txt
streamlit>=1.28.0
requests>=2.31.0
smolagents>=0.0.2
langchain>=0.0.312
pydantic>=2.4.2
openai>=1.2.0
gradio>=5.15.0
pytz>=2023.3
pyyaml>=6.0